# repo/backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Philatelic Museum Operations Suite API
  version: 1.0.0
  description: |
    Public HTTP API for museum operations.

    Authentication model:
    - Session cookie `museum_sid` (set on `POST /auth/login`)
    - `X-CSRF-Token` required on mutating requests (`POST|PUT|PATCH|DELETE`) except `POST /auth/login`
    - `X-Step-Up-Token` required on sensitive actions after `POST /auth/step-up`

    Step-up actions:
    - `GRAPH_PUBLISH`
    - `ROUTE_RULE_CHANGE`
    - `JOB_APPROVE`
    - `JOB_APPEAL_DECIDE`
    - `EXPORT_CREATE`
    - `ADMIN_CONFIG_UPDATE`
servers:
  - url: http://localhost:8080/api/v1
...
security:
  - SessionCookie: []
paths:
...
  /analytics/dimensions:
    post:
      tags: [Analytics]
      summary: Create dimension definition
      description: Requires `ANALYTICS_DIMENSION_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Dimension created }
  /analytics/anomaly-rules:
    post:
      tags: [Analytics]
      summary: Create anomaly rule
      description: Requires `ANALYTICS_METRIC_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Rule created }
  /analytics/dashboards:
    post:
      tags: [Analytics]
      summary: Create dashboard definition
      description: Requires `ANALYTICS_DASHBOARD_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Dashboard created }
  /analytics/dashboards/{dashboardId}:
    get:
      tags: [Analytics]
      summary: Get dashboard values and anomaly evaluations
      description: Requires `ANALYTICS_DASHBOARD_READ`.
      parameters:
        - name: dashboardId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Dashboard payload }
  /analytics/reports:
    post:
      tags: [Analytics]
      summary: Create report definition
      description: Requires `ANALYTICS_REPORT_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      responses:
        '201': { description: Report definition created }
  /analytics/reports/{reportId}/run:
    post:
      tags: [Analytics]
      summary: Trigger report run immediately
      description: Requires `ANALYTICS_REPORT_MANAGE`.
      security:
        - SessionCookie: []
          CsrfHeader: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Report run result }
        '500':
          description: Run failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  /analytics/reports/{reportId}/runs:
    get:
      tags: [Analytics]
      summary: List report runs
      description: Requires `ANALYTICS_REPORT_READ` (Administrator, Auditor).
      parameters:
        - name: reportId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Report runs list }
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: museum_sid
      description: Session cookie set by `POST /auth/login`.
    CsrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required on mutating requests except `POST /auth/login`.
    StepUpHeader:
      type: apiKey
      in: header
      name: X-Step-Up-Token
      description: One-time action-bound step-up token for sensitive operations.
...
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              oneOf:
                - type: array
                  items: { type: object }
                - type: object
            requestId: { type: string }
...